requirements.txt
aiohttp==3.10.0
aiomysql==0.2.0
asgiref==3.8.1
docopt==0.6.2
ipython==8.12.3
matplotlib==3.9.2
nbconvert==7.16.4
pip-chill==1.0.3
python-dotenv==1.0.1
quart==0.19.6
yarg==0.1.9
